Adds Auth Command
