+++ lua/smm/init.lua
@@ -2,11 +2,14 @@ local logger = require 'smm.utils.logger'
 local spotify = require 'smm.spotify'
 local playback = require 'smm.playback'
 local config = require 'smm.config'
+local commands = require 'smm.commands'
 
 local M = {}
 
 function M.setup(user_config)
   config.setup(user_config or {})
+
+  commands.setup()
 
   logger.debug 'Initializing Spotify Module'
   spotify.setup(config.get().spotify)
